+++ src/types/spl/auth.ts
@@ -1,7 +1,7 @@
 export interface SplLoginResponse {
   name: string;
   timestamp: string;
-  token: string;
+  token: string; // this should not be used anymore
   jwt_token: string;
   jwt_expiration_dt: string;
 }